Integerating content progress API with all v4,v5 & v6 version of assessment submit API - KAR-788 (#135)
* Integerating content progress update along with assessment submit API - KAR-788

* Integerating content progress APi with all v4,v5 & v6 version of assessment submit API
